test: 대댓글 작성 API 서비스 테스트 코드 작성
